Got the Menu to render finally
