+++ package.json
@@ -2,9 +2,9 @@
   "name": "sveltekit-tailwind2",
   "version": "0.0.4",
   "scripts": {
-    "dev": "svelte-kit dev",
+    "dev": "svelte-kit dev --port 5000",
     "build": "svelte-kit build",
-    "preview": "svelte-kit preview",
+    "preview": "svelte-kit preview --port 5000",
     "check": "svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
     "format": "prettier --write ./src/**/*.{svelte,ts,js} ./*.{ts,js}",
