+++ src/lib/Componenets/Buttons/ToggleTheme.svelte
@@ -1,0 +1,17 @@
+<script lang="ts">
+  import { theme } from '../../../routes/store'
+
+  const toggleTheme = () => {
+    $theme = $theme === 'dark' ? 'light' : 'dark'
+
+    document.querySelector('html').classList.toggle('dark')
+  }
+</script>
+
+<button on:click={toggleTheme} class="mx-2 text-gray-600 dark:text-gray-300 hover:text-gray-500 dark:hover:text-gray-300"> 
+  <svg class="w-5 h-5 fill-current" viewBox="0 0 24 24" fill="none"
+    xmlns="http://www.w3.org/2000/svg">
+    <path data-name="layer1"
+    d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z"></path>
+  </svg>
+</button>
