+++ src/routes/__layout.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import '../global.postcss'
-  import { theme } from './store'
-  import { onMount } from 'svelte'
-  import MainNav from 'src/components/navigation/MainNav.svelte'
+  import { theme } from './store';
+  import { onMount } from 'svelte';
+  import MainNav from '$lib/Componenets/Navigation/MainNav.svelte';
 
   onMount(() => {
     if ($theme === 'dark') {
@@ -10,8 +10,9 @@
     }
   })
 </script>
-
-<main>
+<header>
   <MainNav></MainNav>
+</header>
+<main>
   <slot />
 </main>
